fixed a bug with eager loading
